feat(app,iam): update TrustaLabs info, remove deprecated lerna call
